Separate file for create total test zip methods
